users can now recieves successful notification mail when registered on the platform
